requirements.txt
asgiref==3.7.2 ; python_version >= "3.11" and python_version < "4.0"
crispy-bootstrap5==0.7 ; python_version >= "3.11" and python_version < "4.0"
django-appconf==1.0.5 ; python_version >= "3.11" and python_version < "4.0"
django-cleanup==7.0.0 ; python_version >= "3.11" and python_version < "4.0"
django-crispy-forms==2.0 ; python_version >= "3.11" and python_version < "4.0"
django-environ==0.10.0 ; python_version >= "3.11" and python_version < "4"
django-imagekit==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
django-tinymce==3.6.1 ; python_version >= "3.11" and python_version < "4.0"
django==4.2.4 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.11" and python_version < "4.0"
pilkit==2.0 ; python_version >= "3.11" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.11" and python_version < "4.0"
psycopg-binary==3.1.10 ; python_version >= "3.11" and python_version < "4.0"
psycopg[binary]==3.1.10 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
pytils==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
setuptools==68.1.2 ; python_version >= "3.11" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.4.4 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
whitenoise==6.5.0 ; python_version >= "3.11" and python_version < "4.0"
